fix: Move UI registry refresh button to layout header
